AppMovil/Legal_AI-main/requirements.txt
pdfplumber
python-docx
streamlit